Nogui argument support and the ability to terminate the server
Added "Gui" parameter (Enables and disables the server interface. Enabled by default.).

Added "Terminate (Kill)" button (Quickly shuts down the server. ATTENTION: This does not save the in-game world and the coordinates of the players.).

When there is no server, the "Create server" button is highlighted in yellow.

When creating the Paper server, if the selected version has already been downloaded, then the previously downloaded version is used instead of re-downloading.